feat(coil-extension): do not auto log in when built for mv3

+++ packages/coil-extension/src/content/services/ContentAuthService.ts
@@ -3,10 +3,13 @@ import { inject, injectable } from 'inversify'
 import { STORAGE_KEY } from '../../types/storage'
 import { ContentRuntime } from '../types/ContentRunTime'
 import * as tokens from '../../types/tokens'
+import { BuildConfig } from '../../types/BuildConfig'
 
 @injectable()
 export class ContentAuthService {
   constructor(
+    @inject(tokens.BuildConfig)
+    private buildConfig: BuildConfig,
     private window: Window,
     private storage: Storage,
     @inject(tokens.ContentRuntime) private contentRuntime: ContentRuntime
@@ -35,22 +38,34 @@ export class ContentAuthService {
       if (token) {
         this.syncViaInjectToken()
       } else {
-        this.contentRuntime.sendMessage({ command: 'logout' })
+        this.sendLogoutMessage()
       }
     })
   }
 
   syncViaInjectToken() {
+    const existingToken = this.storage.getItem(STORAGE_KEY.token)
     this.contentRuntime.sendMessage(
       {
         command: 'injectToken',
-        data: { token: this.storage.getItem(STORAGE_KEY.token) }
+        data: { token: existingToken }
       },
       result => {
         if (result) {
-          this.storage.setItem(STORAGE_KEY.token, result)
+          // We don't want to automatically  login in mv3, only refresh the
+          // page's token with a newer one from the extension. Because in MV3
+          // we are not checking for sign-out on startup.
+          if (this.buildConfig.isMV3 && !existingToken) {
+            this.sendLogoutMessage()
+          } else {
+            this.storage.setItem(STORAGE_KEY.token, result)
+          }
         }
       }
     )
+  }
+
+  private sendLogoutMessage() {
+    this.contentRuntime.sendMessage({ command: 'logout' })
   }
 }
